requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
python-dateutil>=2.8.2 